Calculate viewFinderRect in host view
